+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add highest z-index utility class to Position component
 - Add size modifier for top and bottom paddings to Section component
 - Add `restart` option to Video component
+- Add `hover-target` option to Video component
 
 ### Changed